removed ToasterConfig class in favor of inline config
