--- src/components/Toast/utils/ToasterConfig/index.ts
@@ -1,1 +1,0 @@
-export { ToasterConfig } from "./ToasterConfig"
