+++ src/components/Toast/utils/index.ts
@@ -1,3 +1,2 @@
 export { getToastPositionStyles } from "./getToastPositionStyles"
 export { getToastTransitionStyles } from "./getToastTransitionStyles"
-export { ToasterConfig } from "./ToasterConfig"
